some cleanup based on vercel build errors and root Page.js moved to component to be bounded by suspense
