+++ src/app/components/Timeline.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState, useRef, memo} from "react";
-// import { useNavigate, useLocation } from 'react-router-dom';
 import { FixedSizeList as List } from "react-window";
 import { InView } from "react-intersection-observer";
 import audioIcon from 'public/images/timeline-audio-icon.jpg';
@@ -9,8 +8,6 @@ import zoomInIcon from 'public/images/zoom-in.svg';
 import zoomOutIcon from 'public/images/zoom-out.svg';
 
 const Timeline = ({ items, id, setModalItem, setLocations, setIsLoading, isLoading, setViewContent, setViewMap, setViewTimeline, viewPane }) => {
-    // let navigate = useNavigate();
-    // let location = useLocation();
     const activeItemRef = useRef(null)
     const [activeItemIndex, setActiveItemIndex] = useState(null);
     const [screenSize, setScreenSize] = useState(getCurrentWindowWidth());
